Updated to Akismet 3.0.4 Release Date - 11th December, 2014
Fix to make .htaccess compatible with Apache 2.4.
Fix to allow removal of https author URLs.
Fix to avoid stripping part of the author URL when removing and re-adding.
Removed the "Check for Spam" button from the "Trash" and "Approved" queues, where it would have no effect.
Allow automatic API key configuration when Jetpack is installed and connected to a WordPress.com account